add validation-tools.js
