+++ models/Contact.js
@@ -3,44 +3,25 @@ import Joi from "joi";
 
 import { handleSaveError, runValidatorsAtUpdate } from "./hooks.js";
 
+import { addFieldMongoose, addFieldJoi } from "./validation-tools.js";
+
 
 const contactShemaValidation = {
    name: {
       regExp: /^[a-zA-Zа-яіїєА-ЯІЇЄ]+(([' \-][a-zA-Zа-яіїєА-ЯІЇЄ])?[a-zA-Zа-яіїєА-ЯІЇЄ]*)*$/,
-      errorMessage: "Field name contains invalid characters"
+      errorMessage: "Field name contains invalid characters",
+      requiredErrorMessage: 'Set {PATH} for contact'
    },
    email: {
       regExp: /^([0-9A-Za-z]{1}([-_\.]?[0-9A-Za-z]+)*)@(([0-9A-Za-z]{1}([-_]?[0-9A-Za-z]+)*\.){1,2}[A-Za-z]{2,})$/,
-      errorMessage: "Field email must be a valid email"
+      errorMessage: "Field email must be a valid email",
+      requiredErrorMessage: 'Set {PATH} for contact'
    },
     phone: {
       regExp: /^\(\d{3}\) \d{3}-\d{4}$/,
-      errorMessage: "Field phone must be in the format (XXX) XXX-XXXX"
+      errorMessage: "Field phone must be in the format (XXX) XXX-XXXX",
+      requiredErrorMessage: 'Set {PATH} for contact'
    }
-}
-
-const addFieldMongoose = fieldName => {
-   const {regExp, errorMessage} = contactShemaValidation[fieldName]
-   return {
-      type: String,
-      match: [regExp, errorMessage],
-      required: [true, 'Set {PATH} for contact'],
-   }
-}
-
-const messagesErrorsJoi = message => {
-   return {
-      "string.empty": "missing required {#label} field", 
-      "any.required": "missing required {#label} field",
-      "string.pattern.base": message
-   }
-}
-
-function addFieldJoi( fieldName) {
-   const {regExp, errorMessage} = contactShemaValidation[fieldName]
-   return this.string().required()
-      .pattern(new RegExp(regExp))
-        .messages(messagesErrorsJoi(errorMessage))
 }
 
 const messagesFavoriteErrors = {
@@ -51,9 +32,9 @@ const messagesFavoriteErrors = {
 
 // Mongoose
 const contactSchemaDB = new Schema({
-   name: { ...addFieldMongoose("name"), minlength: 3 },
-   email: addFieldMongoose("email"),
-   phone: addFieldMongoose("phone"),
+   name: { ...addFieldMongoose(contactShemaValidation.name), minlength: 3 },
+   email: addFieldMongoose(contactShemaValidation.email),
+   phone: addFieldMongoose(contactShemaValidation.phone),
    favorite: {
     type: Boolean,
     default: false,
@@ -62,8 +43,8 @@ const contactSchemaDB = new Schema({
       type: Schema.Types.ObjectId,
       ref: 'user',
     }
-   },  { versionKey: false, timestamps: true })
-
+}, { versionKey: false, timestamps: true })
+ 
 contactSchemaDB.post("save", handleSaveError);
 
 contactSchemaDB.pre("findOneAndUpdate", runValidatorsAtUpdate);
@@ -75,10 +56,10 @@ const Contact = model("contact", contactSchemaDB);
 
 // Joi
 export const contactSchemaAll = Joi.object({
-   name: addFieldJoi.call(Joi, "name")
+   name: addFieldJoi.call(Joi, contactShemaValidation.name)
          .min(3),
-    email: addFieldJoi.call(Joi,"email"),
-    phone: addFieldJoi.call(Joi,"phone"),
+    email: addFieldJoi.call(Joi, contactShemaValidation.email),
+    phone: addFieldJoi.call(Joi, contactShemaValidation.phone),
     favorite: Joi.bool()
 })
